small bugfixes, and kill other warning
git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@4466 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ Gui/mplayer/menu.h
@@ -11,7 +11,7 @@ void mplMenuDraw( wsParamDisplay )
 {
  unsigned long * buf = NULL;
  unsigned long * drw = NULL;
- unsigned long   x,y,tmp;
+ int             x,y,tmp;
 
  if ( !appMPlayer.menuBase.Bitmap.Image ) return;
  if ( !mplMenuRender && !appMPlayer.menuWindow.Visible ) return;
